requirements.txt
anthropic>=0.40.0
trueskill>=0.4.5
openai>=1.0.0
google-generativeai>=0.8.0
fastapi>=0.115.0
uvicorn>=0.30.0
pytest>=8.0.0
httpx>=0.27.0
